finalize the genetic algorithm

+++ run_ga.py
@@ -61,8 +61,8 @@ def run_ga():
             # p2_ind = poplib.Population.select_parent(fitmap)
 
             # tournament style selection
-            p1_ind = poplib.Population.select_parent_tournament(fitness_scores, tournament_size=POPULATION/5)
-            p2_ind = poplib.Population.select_parent_tournament(fitness_scores, tournament_size=POPULATION/5)
+            p1_ind = poplib.Population.select_parent_tournament(fitness_scores, tournament_size=int(POPULATION/5))
+            p2_ind = poplib.Population.select_parent_tournament(fitness_scores, tournament_size=int(POPULATION/5))
 
             dna = genlib.Genome.crossover(pop.creatures[p1_ind].dna,
                                         pop.creatures[p2_ind].dna)
